#3-1 [Add] 약관 사용자 연결 구현

+++ backend/src/main/java/dragonb/bearfamily/backend/repository/TermsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dragonb.bearfamily.backend.model.Terms;
 
-public interface TermsRepository extends JpaRepository<Terms, Integer>{
+public interface TermsRepository extends JpaRepository<Terms, Long>{
     
 }
